cast

+++ src/kex.c
@@ -1540,7 +1540,7 @@ kex_method_diffie_hellman_group_exchange_sha1_key_exchange(
             goto dh_gex_clean_exit;
         }
 
-        bits = _libssh2_bn_bits(key_state->p);
+        bits = (int)_libssh2_bn_bits(key_state->p);
         if(bits < 0 || bits < LIBSSH2_DH_GEX_MINGROUP ||
            bits > LIBSSH2_DH_GEX_MAXGROUP) {
             ret = _libssh2_error(session, LIBSSH2_ERROR_PROTO,
@@ -1669,7 +1669,7 @@ kex_method_diffie_hellman_group_exchange_sha256_key_exchange(
             goto dh_gex_clean_exit;
         }
 
-        bits = _libssh2_bn_bits(key_state->p);
+        bits = (int)_libssh2_bn_bits(key_state->p);
         if(bits < 0 || bits < LIBSSH2_DH_GEX_MINGROUP ||
            bits > LIBSSH2_DH_GEX_MAXGROUP) {
             ret = _libssh2_error(session, LIBSSH2_ERROR_PROTO,
